requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.1.0
